add highlighting for result
the scope is definitely open to debate. i chose support.class because it was already assigned to a mostly unused color (except for escaped chars) for nim in the theme i use, and it may not be the most perfect and fitting scope of all that are commonly used. having result highlighted (and highlighted differently) is def personal preference but given that has additional meaning compared to some user declared variable, and could be used frequently, it felt right to have it highlighted. it would be reasonable to have it the same color as return, but since it works a bit differently than the keywords i thought it would be fitting to make it different. additionally it could have a different regex that only matched it in a proc or func or maybe a macro/template